test: move SPI tests into spi package
